Create new metadata object

You can create a new metadata by specifying each of the elements, by specifying a list, or by providing a dataset which includes metadata.

stw_meta(...)

# S3 method for default
stw_meta(...)

# S3 method for character
stw_meta(
  name,
  title = NULL,
  description = NULL,
  sources = NULL,
  n_row = NULL,
  n_col = NULL,
  dict = NULL,
  ...
)

# S3 method for list
stw_meta(env, ...)

# S3 method for stw_meta
stw_meta(meta, ...)

# S3 method for stw_dataset
stw_meta(dataset, ...)

Arguments

...
additional args (not used)

name
character name of the dataset

title
character title of the dataset

description
character description of the dataset

sources
list source of the dataset; has elements title, and optionally path and email

n_row
integer number of rows in the dataset

n_col
integer number of columns in the dataset

dict
Object with S3 class stw_dict, contains data-dictionary

env
list with elements name, title, etc.

meta
Object with S3 class stw_meta, contains meta-data

dataset
Object with S3 class stw_dataset, a data frame with attached meta-data

Value

Object with S3 class stw_meta

Details

When you create a metadata object, it is checked for required and optional elements. Each metadaset is required to have a name and dict object. A title, description, source, n_row, n_col are optional.